[REFACTOR] 리팩토링
