--- operation-api/src/main/java/org/sopt/makers/operation/common/scheduler/service/LectureService.java
@@ -1,6 +1,0 @@
-package org.sopt.makers.operation.common.scheduler.service;
-
-public interface LectureService {
-
-	void endLectures();
-}
