--- operation-api/src/main/java/org/sopt/makers/operation/service/app/member/dto/response/MemberScoreGetResponse.java
@@ -1,11 +1,0 @@
-package org.sopt.makers.operation.service.app.member.dto.response;
-
-public record MemberScoreGetResponse(
-		float score
-) {
-	public static MemberScoreGetResponse of(float score){
-		return new MemberScoreGetResponse(
-				score
-		);
-	}
-}
