--- operation-api/src/main/java/org/sopt/makers/operation/service/web/admin/dto/request/LoginRequest.java
@@ -1,7 +1,0 @@
-package org.sopt.makers.operation.service.web.admin.dto.request;
-
-public record LoginRequest(
-		String email,
-		String password
-) {
-}
