--- operation-api/src/main/java/org/sopt/makers/operation/service/web/admin/dto/request/SignUpRequest.java
@@ -1,21 +1,0 @@
-package org.sopt.makers.operation.service.web.admin.dto.request;
-
-import org.operation.admin.domain.Admin;
-import org.operation.admin.domain.Role;
-
-public record SignUpRequest(
-        String email,
-        String password,
-        String name,
-        Role role
-) {
-
-	public Admin toEntity() {
-		return Admin.builder()
-				.email(this.email)
-				.password(this.password)
-				.name(this.name)
-				.role(this.role)
-				.build();
-	}
-}
