--- operation-api/src/main/java/org/sopt/makers/operation/service/web/admin/dto/response/RefreshResponse.java
@@ -1,13 +1,0 @@
-package org.sopt.makers.operation.service.web.admin.dto.response;
-
-import org.operation.admin.domain.Admin;
-
-public record RefreshResponse(
-        String accessToken,
-        String refreshToken
-) {
-
-    public static RefreshResponse of(String accessToken, Admin admin) {
-        return new RefreshResponse(accessToken, admin.getRefreshToken());
-    }
-}
