--- operation-api/src/main/java/org/sopt/makers/operation/service/web/admin/repository/AdminRepository.java
@@ -1,14 +1,0 @@
-package org.sopt.makers.operation.service.web.admin.repository;
-
-import java.util.Optional;
-
-import org.operation.admin.domain.Admin;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
-    Optional<Admin> findByEmail(String email);
-    boolean existsByEmail(String email);
-
-}
